+++ presentation/src/main/java/com/danbam/presentation/ui/find/find_password/FindPasswordScreen.kt
@@ -1,4 +1,4 @@
-package com.danbam.presentation.ui.find
+package com.danbam.presentation.ui.find.find_password
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -10,6 +10,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.danbam.design_system.component.HeadLineBold
@@ -20,11 +21,14 @@ import com.danbam.presentation.R
 import com.danbam.presentation.util.AppNavigationItem
 
 @Composable
-fun FindIdScreen(
+fun FindPasswordScreen(
     navController: NavController,
     phoneNumber: String,
 ) {
-    var currentId by remember { mutableStateOf("") }
+    var password by remember { mutableStateOf("") }
+    var passwordVisible by remember { mutableStateOf(false) }
+    var checkPassword by remember { mutableStateOf("") }
+    var checkPasswordVisible by remember { mutableStateOf(false) }
 
     Column(
         modifier = Modifier.fillMaxSize()
@@ -35,17 +39,27 @@ fun FindIdScreen(
         HeadLineBold(
             modifier = Modifier
                 .padding(start = 32.dp, top = 16.dp),
-            text = stringResource(id = R.string.current_id)
+            text = stringResource(id = R.string.change_password)
         )
         IndiStrawTextField(
-            modifier = Modifier.padding(top = 96.dp),
-            hint = "",
-            value = currentId,
-            onValueChange = { },
-            readOnly = true
+            modifier = Modifier.padding(top = 66.dp),
+            hint = stringResource(id = R.string.password),
+            value = password,
+            onValueChange = { password = it },
+            imeAction = ImeAction.Next,
+            isToggleVisible = passwordVisible,
+            onToggleChange = { passwordVisible = !passwordVisible }
+        )
+        IndiStrawTextField(
+            modifier = Modifier.padding(top = 20.dp),
+            hint = stringResource(id = R.string.check_password),
+            value = checkPassword,
+            onValueChange = { checkPassword = it },
+            isToggleVisible = checkPasswordVisible,
+            onToggleChange = { checkPasswordVisible = !checkPasswordVisible }
         )
         IndiStrawButton(
-            modifier = Modifier.padding(top = 78.dp),
+            modifier = Modifier.padding(top = 37.dp),
             text = stringResource(id = R.string.check)
         ) {
             navController.navigate(AppNavigationItem.Login.route) {
